Stripped configuration out into separate classes
