+++ Commands/GreetCommand.cs
@@ -7,24 +7,6 @@ namespace ConsoleStarter.Commands
     public class GreetCommand : ICommand
     {
 
-        public static void Configure(CommandLineApplication command, CommandLineOptions options)
-        {
-
-            command.Description = "An example command from the neat .NET Core Starter";
-            command.HelpOption("--help|-h|-?");
-
-            var nameArgument = command.Argument("name",
-                                   "Name I should say hello to");
-
-            command.OnExecute(() =>
-                {
-                    options.Command = new GreetCommand(nameArgument.Value, options);
-
-                    return 0;
-                });
-
-        }
-
         private readonly string _name;
         private readonly CommandLineOptions _options;
 
@@ -34,11 +16,13 @@ namespace ConsoleStarter.Commands
             _options = options;
         }
 
-        public void Run()
+        public int Run()
         {
             Console.WriteLine("Hello "
                 + (_name != null ? _name : "World")
                 + (_options.IsEnthousiastic ? "!!!" : "."));
+
+            return 0;
         }
 
     }
