+++ Commands/ICommand.cs
@@ -3,7 +3,7 @@ namespace ConsoleStarter
 
     public interface ICommand
     {
-        void Run();
+        int Run();
     }
 
 }
