+++ Commands/RootCommand.cs
@@ -5,21 +5,6 @@ namespace ConsoleStarter.Commands
 
     public class RootCommand : ICommand
     {
-
-        public static void Configure(CommandLineApplication app, CommandLineOptions options)
-        {
-
-            app.Command("greet", c => GreetCommand.Configure(c, options));
-
-            app.OnExecute(() =>
-                {
-                    options.Command = new RootCommand(app);
-
-                    return 0;
-                });
-
-        }
-
         private readonly CommandLineApplication _app;
 
         public RootCommand(CommandLineApplication app)
@@ -27,9 +12,11 @@ namespace ConsoleStarter.Commands
             _app = app;
         }
 
-        public void Run()
+        public int Run()
         {
             _app.ShowHelp();
+
+            return 1;
         }
 
     }
